add attributes in symbol table and fix Round

+++ src/codgen/reducers/Sum.java
@@ -5,12 +5,12 @@ import java.util.ArrayList;
 public class Sum implements AggregationFunction {
     @Override
     public String reduce(ArrayList<String> values) {
-        Integer sum = 0;
-        ArrayList<Integer> integerValues = new ArrayList<>();
+        Double sum = 0.0;
+        ArrayList<Double> integerValues = new ArrayList<>();
 
-        values.forEach(e->integerValues.add(Integer.parseInt(e)));
+        values.forEach(e->integerValues.add(Double.parseDouble(e)));
 
-        for(Integer value : integerValues)
+        for(Double value : integerValues)
             sum+=value;
 
         return String.valueOf(sum);
